fix on/off
https://github.com/tgstation/tgstation/pull/72341

+++ code/game/objects/items/implants/implantuplink.dm
@@ -24,6 +24,9 @@ CREATION_TEST_IGNORE_SUBTYPES(/obj/item/implant/uplink)
  * the component, so delete itself.
  */
 /obj/item/implant/uplink/proc/_component_removal(datum/source, datum/component/component)
+	SIGNAL_HANDLER
+	if(QDELETED(src))
+		return
 	if(istype(component, /datum/component/uplink))
 		qdel(src)
